feat(build): upgrade werker build and vendored dependencies

+++ scripts/test.sh
@@ -1,5 +1,4 @@
 #!/bin/sh
-#go test -v --race $(find . -maxdepth 10 -not -path './.git*' -not -path '*/_*' -type d | grep -v examples | grep -v pkg | grep -v bin | grep -v scripts  | xargs -I{} echo "{}/")
 
 # Get Test dependencies
 go get github.com/axw/gocov/gocov
@@ -8,12 +7,18 @@ go get golang.org/x/tools/cmd/cover
 go get github.com/modocache/gover
 
 # Run test coverage on each subdirectories and merge the coverage profile.
+
 echo "mode: count" > profile.cov
 
 # Standard go tooling behavior is to ignore dirs with leading underscors
-for dir in $(find . -maxdepth 10 -not -path './.git*' -not -path '*/_*' -type d | grep -v examples); do
+for dir in $(find . -maxdepth 3 -not -path './.git*' -not -path '*/_*' -not -path '_*' -type d -not -path './examples*' -not -path './vendor*' -not -path './output*' -not -path './dist*' -not -path './build*' -not -path './bin*' -not -path './scripts*' -not -path './pact-*'); do
   if ls $dir/*.go &> /dev/null; then
     go test -covermode=count -coverprofile=$dir/profile.tmp $dir
+    if [ $? = 1 ]; then
+      echo "Test failure, exiting"
+      exit 1
+    fi
+
     if [ -f $dir/profile.tmp ]; then
     	cat $dir/profile.tmp | tail -n +2 >> profile.cov
     	rm $dir/profile.tmp
